# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/jenis-perizinan": {
            "get": {
                "tags": [
                    "Jenis Perizinan"
                ],
                "summary": "Get all jenis perizinan",
                "description": "Returns all jenis perizinan data",
                "operationId": "getAllJenisPerizinan",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DataResponse"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Jenis Perizinan"
                ],
                "summary": "Create new jenis perizinan",
                "description": "Create a new jenis perizinan data",
                "operationId": "createJenisPerizinan",
                "requestBody": {
                    "description": "Jenis Perizinan data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JenisPerizinanInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jenis Perizinan created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jenis-perizinan/{id}": {
            "get": {
                "tags": [
                    "Jenis Perizinan"
                ],
                "summary": "Get jenis perizinan by ID",
                "description": "Returns a single jenis perizinan data",
                "operationId": "getJenisPerizinan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Jenis Perizinan ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Jenis Perizinan not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Jenis Perizinan"
                ],
                "summary": "Update jenis perizinan by ID",
                "description": "Update a single jenis perizinan data",
                "operationId": "updateJenisPerizinan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Jenis Perizinan ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Jenis Perizinan data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JenisPerizinanInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jenis Perizinan updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Jenis Perizinan"
                ],
                "summary": "Delete by ID",
                "operationId": "deleteJenisPerizinan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Jenis Perizinan record deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Jenis Perizinan record not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "DataResponse": {
                "description": "Data response dari permintaan perizinan",
                "properties": {
                    "status": {
                        "description": "Keterangan status response",
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "description": "Pesan terkait response",
                        "type": "string",
                        "example": "Permintaan perizinan berhasil diajukan"
                    },
                    "data": {
                        "description": "Data API",
                        "type": "object"
                    }
                },
                "type": "object",
                "example": {
                    "status": "success",
                    "message": "Permintaan perizinan berhasil diajukan",
                    "data": {
                        "nama": "John Doe",
                        "alamat": "Jl. Kenanga No. 123",
                        "jenis_perizinan": "Izin Usaha",
                        "tanggal_mulai": "2023-04-01",
                        "tanggal_selesai": "2023-06-30"
                    }
                }
            },
...
            "JenisPerizinanInput": {
                "title": "Jenis Perizinan Input",
                "required": [
                    "nama_jenis",
                    "deskripsi"
                ],
                "properties": {
                    "nama_jenis": {
                        "description": "Nama jenis perizinan",
                        "type": "string"
                    },
                    "deskripsi_perizinan": {
                        "description": "Deskripsi jenis perizinan",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}